@@ -11,26 +11,25 @@
     "@mui/material": "^5.18.0",
     "@mui/styles": "^5.18.0",
     "@mui/x-date-pickers": "^8.21.0",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "@vitejs/plugin-react": "^5.1.1",
     "axios": "^1.13.2",
     "dayjs": "^1.11.0",
-    "i18next": "^21.6.14",
-    "react": "^17.0.2",
+    "i18next": "^23.11.2",
+    "react": "^18.3.1",
     "react-csv": "^2.2.2",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.15.7",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
+    "react-dom": "^18.3.1",
+    "react-i18next": "^14.0.0",
+    "react-router": "^6.30.2",
+    "react-router-dom": "^6.30.2",
     "recharts": "^2.1.9",
     "styled-components": "^5.3.3",
     "vite": "^7.2.6"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.1",
-    "@types/react-router-dom": "^5.3.3",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^15.0.7",
+    "@testing-library/user-event": "^14.5.2",
+    "@vitejs/plugin-react": "^5.1.1",
     "env-cmd": "^10.1.0",
     "eslint": "^8.57.1",
     "eslint-config-prettier": "^8.1.0",
