@@ -1,5 +1,5 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import App from './App'
 import './index.css'
 import * as serviceWorker from './serviceWorker'
@@ -9,17 +9,16 @@ const props = {}
 
 if (root) {
   const attrs = Object.keys(root.dataset)
-  attrs.forEach(
-    (name) => { props[name] = root.dataset[name] }
-  )
+  attrs.forEach((name) => {
+    props[name] = root.dataset[name]
+  })
 
   console.log(`infoenergia version: ${import.meta.env.VITE_APP_VERSION}`)
 
-  ReactDOM.render(
-    <>
+  ReactDOM.createRoot(document.getElementById('root')).render(
+    <React.StrictMode>
       <App {...props} />
-    </>,
-    document.getElementById('root')
+    </React.StrictMode>
   )
 }
 
@@ -27,4 +26,3 @@ if (root) {
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
 serviceWorker.unregister()
-
